Move NameGenerator to quasar.effect
