+++ connector/src/main/scala/quasar/qscript/qsu/ExpandShifts.scala
@@ -18,7 +18,7 @@ package quasar.qscript.qsu
 
 import slamdata.Predef._
 
-import quasar.NameGenerator
+import quasar.effect.NameGenerator
 import quasar.Planner.PlannerErrorME
 import quasar.contrib.scalaz._
 import quasar.ejson.EJson
