+++ connector/src/main/scala/quasar/qscript/qsu/ExtractFreeMap.scala
@@ -18,8 +18,9 @@ package quasar.qscript.qsu
 
 import slamdata.Predef.{Map => SMap, _}
 
-import quasar.{NameGenerator, RenderTreeT}
+import quasar.RenderTreeT
 import quasar.Planner.{InternalError, PlannerErrorME}
+import quasar.effect.NameGenerator
 import quasar.fp.symbolOrder
 import quasar.qscript.RecFreeS._
 import quasar.qscript.{construction, JoinSide, LeftSide, RightSide}
