+++ connector/src/main/scala/quasar/qscript/qsu/Graduate.scala
@@ -18,7 +18,7 @@ package quasar.qscript.qsu
 
 import slamdata.Predef._
 
-import quasar.NameGenerator
+import quasar.effect.NameGenerator
 import quasar.Planner.{InternalError, PlannerErrorME}
 import quasar.common.JoinType
 import quasar.contrib.pathy.AFile
