+++ connector/src/main/scala/quasar/qscript/qsu/minimizers/Minimizer.scala
@@ -17,8 +17,9 @@
 package quasar.qscript.qsu
 package minimizers
 
-import quasar.{NameGenerator, Planner}, Planner.PlannerErrorME
 import slamdata.Predef._
+import quasar.Planner, Planner.PlannerErrorME
+import quasar.effect.NameGenerator
 
 import scalaz.Monad
