+++ mongodb/src/main/scala/quasar/physical/mongodb/fs/QueryFileInterpreter.scala
@@ -22,6 +22,7 @@ import quasar.common.{PhaseResult, PhaseResults, PhaseResultT}
 import quasar.contrib.pathy._
 import quasar.contrib.scalaz.eitherT._
 import quasar.contrib.scalaz.kleisli._
+import quasar.effect.NameGenerator
 import quasar.fp._
 import quasar.fp.ski._
 import quasar.frontend.logicalplan.{constant, LogicalPlan}
